Move template specialization of RTreeInfo for MultidimensionalRange
