+++ lib/Modeling/IndexSet.cpp
@@ -6,23 +6,6 @@
 
 using namespace ::marco::modeling;
 
-namespace marco::modeling {
-const MultidimensionalRange &
-RTreeInfo<MultidimensionalRange>::getShape(const MultidimensionalRange &val) {
-  return val;
-}
-
-bool RTreeInfo<MultidimensionalRange>::isEqual(
-    const MultidimensionalRange &first, const MultidimensionalRange &second) {
-  return first == second;
-}
-
-void RTreeInfo<MultidimensionalRange>::dump(llvm::raw_ostream &os,
-                                            const MultidimensionalRange &val) {
-  os << val;
-}
-} // namespace marco::modeling
-
 namespace {
 template <typename>
 struct Merger {};
